pyproject.toml
[project]
name = "blive-dglab"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.104.1",
    "uvicorn==0.24.0",
    "bilibili-api-python>=9.0.0",
    "qrcode==7.4.2",
    "pillow==10.1.0",
    "jinja2==3.1.2",
    "python-multipart==0.0.6",
]

[[tool.uv.index]]
name = "tsinghua-source"
url = "https://pypi.tuna.tsinghua.edu.cn/simple"